[Ref] - Add some refactoring in requirement and folders titles
