+++ 4task/wagons/cargo_wagon.rb
